fix msvc build warnings and errors
This commit fix the msvc2013 build for configuration:
configure --target=x86_64-win64-vs12 --enable-experimental
 --enable-clpf --enable-dering --enable-motion-var --enable-ans

BUG=aomedia:80

+++ aom_dsp/bitreader.h
@@ -74,7 +74,7 @@ static INLINE int aom_reader_init(aom_reader *r, const uint8_t *buffer,
   (void)decrypt_cb;
   (void)decrypt_state;
   assert(size <= INT_MAX);
-  return ans_read_init(r, buffer, size);
+  return ans_read_init(r, buffer, (int)size);
 #elif CONFIG_DAALA_EC
   (void)decrypt_cb;
   (void)decrypt_state;
